tests/wpt/web-platform-tests/html/webappapis/dynamic-markup-insertion/closing-the-input-stream/document.close-01.xhtml

<html xmlns="http://www.w3.org/1999/xhtml">
<head>
<title>document.close in XHTML</title>
<link rel="author" title="Ms2ger" href="mailto:ms2ger@gmail.com"/>
<link rel="help" href="https://html.spec.whatwg.org/multipage/#closing-the-input-stream"/>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
</head>
<body>
<div id="log"></div>
<script>
test(function() {
  assert_throws("INVALID_STATE_ERR", function() {
    document.close();
  }, "document.close in XHTML should throw an INVALID_STATE_ERR ");
}, "document.close in XHTML");
</script>
</body>
</html>
